helpmenuを分割する。 #51
