+++ docs/html/_modules/index.html
@@ -88,7 +88,7 @@
 <li class="toctree-l1"><a class="reference internal" href="../sub/save.html">保存ファイルについて</a></li>
 <li class="toctree-l1"><a class="reference internal" href="../sub/exe.html">exe実行ファイルを作成</a></li>
 <li class="toctree-l1"><a class="reference internal" href="../sub/coverage.html">カバレッジの計測について</a></li>
-<li class="toctree-l1"><a class="reference internal" href="../novel_editor.html">Novel Editor モジュール</a></li>
+<li class="toctree-l1"><a class="reference internal" href="../modules.html">source</a></li>
 </ul>
 
             
@@ -151,7 +151,10 @@
            <div itemprop="articleBody">
             
   <h1>全モジュールのうち、コードを読めるもの</h1>
-<ul><li><a href="novel_editor.html">novel_editor</a></li>
+<ul><li><a href="helpmenu.html">helpmenu</a></li>
+<li><a href="mydialog.html">mydialog</a></li>
+<li><a href="novel_editor.html">novel_editor</a></li>
+<li><a href="processingmenu.html">processingmenu</a></li>
 </ul>
 
            </div>
