+++ docs/html/_sources/helpmenu.rst.txt
@@ -1,0 +1,7 @@
+helpmenu module
+===============
+
+.. automodule:: helpmenu
+   :members:
+   :undoc-members:
+   :show-inheritance:
