+++ docs/html/_sources/mydialog.rst.txt
@@ -1,0 +1,7 @@
+mydialog module
+===============
+
+.. automodule:: mydialog
+   :members:
+   :undoc-members:
+   :show-inheritance:
