+++ docs/html/_sources/processingmenu.rst.txt
@@ -1,0 +1,7 @@
+processingmenu module
+=====================
+
+.. automodule:: processingmenu
+   :members:
+   :undoc-members:
+   :show-inheritance:
