+++ docs/html/genindex.html
@@ -88,7 +88,7 @@
 <li class="toctree-l1"><a class="reference internal" href="sub/save.html">保存ファイルについて</a></li>
 <li class="toctree-l1"><a class="reference internal" href="sub/exe.html">exe実行ファイルを作成</a></li>
 <li class="toctree-l1"><a class="reference internal" href="sub/coverage.html">カバレッジの計測について</a></li>
-<li class="toctree-l1"><a class="reference internal" href="novel_editor.html">Novel Editor モジュール</a></li>
+<li class="toctree-l1"><a class="reference internal" href="modules.html">source</a></li>
 </ul>
 
             
@@ -162,6 +162,7 @@
  | <a href="#E"><strong>E</strong></a>
  | <a href="#F"><strong>F</strong></a>
  | <a href="#G"><strong>G</strong></a>
+ | <a href="#H"><strong>H</strong></a>
  | <a href="#I"><strong>I</strong></a>
  | <a href="#L"><strong>L</strong></a>
  | <a href="#M"><strong>M</strong></a>
@@ -172,6 +173,7 @@
  | <a href="#S"><strong>S</strong></a>
  | <a href="#T"><strong>T</strong></a>
  | <a href="#U"><strong>U</strong></a>
+ | <a href="#V"><strong>V</strong></a>
  | <a href="#Y"><strong>Y</strong></a>
  | <a href="#モ"><strong>モ</strong></a>
  
@@ -241,6 +243,8 @@
 </li>
       <li><a href="novel_editor.html#novel_editor.LineFrame.find_dictionaly">find_dictionaly() (novel_editor.LineFrame のメソッド)</a>
 </li>
+      <li><a href="processingmenu.html#processingmenu.ProcessingMenuClass.find_wikipedia">find_wikipedia() (processingmenu.ProcessingMenuClass のメソッド)</a>
+</li>
       <li><a href="novel_editor.html#novel_editor.LineFrame.font_dialog">font_dialog() (novel_editor.LineFrame のメソッド)</a>
 </li>
   </ul></td>
@@ -264,6 +268,25 @@
   </ul></td>
   <td style="width: 33%; vertical-align: top;"><ul>
       <li><a href="novel_editor.html#novel_editor.LineFrame.get_keywords">get_keywords() (novel_editor.LineFrame のメソッド)</a>
+</li>
+  </ul></td>
+</tr></table>
+
+<h2 id="H">H</h2>
+<table style="width: 100%" class="indextable genindextable"><tr>
+  <td style="width: 33%; vertical-align: top;"><ul>
+      <li><a href="helpmenu.html#helpmenu.HelpMenuClass.help">help() (helpmenu.HelpMenuClass のメソッド)</a>
+</li>
+      <li>
+    helpmenu
+
+      <ul>
+        <li><a href="helpmenu.html#module-helpmenu">モジュール</a>
+</li>
+      </ul></li>
+  </ul></td>
+  <td style="width: 33%; vertical-align: top;"><ul>
+      <li><a href="helpmenu.html#helpmenu.HelpMenuClass">HelpMenuClass (helpmenu のクラス)</a>
 </li>
   </ul></td>
 </tr></table>
@@ -299,13 +322,20 @@
 </li>
       <li><a href="novel_editor.html#novel_editor.LineFrame.moji_count">moji_count() (novel_editor.LineFrame のメソッド)</a>
 </li>
-  </ul></td>
-  <td style="width: 33%; vertical-align: top;"><ul>
       <li><a href="novel_editor.html#novel_editor.LineFrame.mouse_image_scroll">mouse_image_scroll() (novel_editor.LineFrame のメソッド)</a>
 </li>
+  </ul></td>
+  <td style="width: 33%; vertical-align: top;"><ul>
       <li><a href="novel_editor.html#novel_editor.LineFrame.mouse_y_scroll">mouse_y_scroll() (novel_editor.LineFrame のメソッド)</a>
 </li>
-      <li><a href="novel_editor.html#novel_editor.Mydialog">Mydialog (novel_editor のクラス)</a>
+      <li>
+    mydialog
+
+      <ul>
+        <li><a href="mydialog.html#module-mydialog">モジュール</a>
+</li>
+      </ul></li>
+      <li><a href="mydialog.html#mydialog.Mydialog">Mydialog (mydialog のクラス)</a>
 </li>
   </ul></td>
 </tr></table>
@@ -340,10 +370,12 @@
 </li>
       <li><a href="novel_editor.html#novel_editor.LineFrame.on_name_click">on_name_click() (novel_editor.LineFrame のメソッド)</a>
 </li>
-      <li><a href="novel_editor.html#novel_editor.LineFrame.open_file">open_file() (novel_editor.LineFrame のメソッド)</a>
+      <li><a href="processingmenu.html#processingmenu.ProcessingMenuClass.open_becoming_novelist_page">open_becoming_novelist_page() (processingmenu.ProcessingMenuClass のメソッド)</a>
 </li>
   </ul></td>
   <td style="width: 33%; vertical-align: top;"><ul>
+      <li><a href="novel_editor.html#novel_editor.LineFrame.open_file">open_file() (novel_editor.LineFrame のメソッド)</a>
+</li>
       <li><a href="novel_editor.html#novel_editor.LineFrame.open_file_save">open_file_save() (novel_editor.LineFrame のメソッド)</a>
 </li>
       <li><a href="novel_editor.html#novel_editor.LineFrame.open_help">open_help() (novel_editor.LineFrame のメソッド)</a>
@@ -368,8 +400,17 @@
 </li>
       <li><a href="novel_editor.html#novel_editor.LineFrame.print_gif">print_gif() (novel_editor.LineFrame のメソッド)</a>
 </li>
+      <li>
+    processingmenu
+
+      <ul>
+        <li><a href="processingmenu.html#module-processingmenu">モジュール</a>
+</li>
+      </ul></li>
   </ul></td>
   <td style="width: 33%; vertical-align: top;"><ul>
+      <li><a href="processingmenu.html#processingmenu.ProcessingMenuClass">ProcessingMenuClass (processingmenu のクラス)</a>
+</li>
       <li><a href="novel_editor.html#novel_editor.LineFrame.push_keys">push_keys() (novel_editor.LineFrame のメソッド)</a>
 </li>
       <li><a href="novel_editor.html#novel_editor.LineFrame.pyttsx3_onend">pyttsx3_onend() (novel_editor.LineFrame のメソッド)</a>
@@ -422,9 +463,9 @@
 </li>
       <li><a href="novel_editor.html#novel_editor.LineFrame.selection">selection() (novel_editor.LineFrame のメソッド)</a>
 </li>
-      <li><a href="novel_editor.html#novel_editor.Mydialog.sub_name_ok">sub_name_ok() (novel_editor.Mydialog のメソッド)</a>
+      <li><a href="mydialog.html#mydialog.Mydialog.sub_name_ok">sub_name_ok() (mydialog.Mydialog のメソッド)</a>
 </li>
-      <li><a href="novel_editor.html#novel_editor.Mydialog.self.sub_name_win">sub_name_win (novel_editor.Mydialog.self の属性)</a>
+      <li><a href="mydialog.html#mydialog.Mydialog.self.sub_name_win">sub_name_win (mydialog.Mydialog.self の属性)</a>
 </li>
   </ul></td>
 </tr></table>
@@ -438,9 +479,9 @@
 </li>
   </ul></td>
   <td style="width: 33%; vertical-align: top;"><ul>
-      <li><a href="novel_editor.html#novel_editor.Mydialog.self.txt">txt (novel_editor.Mydialog.self の属性)</a>
+      <li><a href="mydialog.html#mydialog.Mydialog.self.txt">txt (mydialog.Mydialog.self の属性)</a>
 </li>
-      <li><a href="novel_editor.html#novel_editor.Mydialog.self.txt_name">txt_name (novel_editor.Mydialog.self の属性)</a>
+      <li><a href="mydialog.html#mydialog.Mydialog.self.txt_name">txt_name (mydialog.Mydialog.self の属性)</a>
 </li>
   </ul></td>
 </tr></table>
@@ -453,6 +494,14 @@
   </ul></td>
   <td style="width: 33%; vertical-align: top;"><ul>
       <li><a href="novel_editor.html#novel_editor.LineFrame.update_line_numbers">update_line_numbers() (novel_editor.LineFrame のメソッド)</a>
+</li>
+  </ul></td>
+</tr></table>
+
+<h2 id="V">V</h2>
+<table style="width: 100%" class="indextable genindextable"><tr>
+  <td style="width: 33%; vertical-align: top;"><ul>
+      <li><a href="helpmenu.html#helpmenu.HelpMenuClass.version">version() (helpmenu.HelpMenuClass のメソッド)</a>
 </li>
   </ul></td>
 </tr></table>
@@ -478,7 +527,13 @@
     モジュール
 
       <ul>
+        <li><a href="helpmenu.html#module-helpmenu">helpmenu</a>
+</li>
+        <li><a href="mydialog.html#module-mydialog">mydialog</a>
+</li>
         <li><a href="novel_editor.html#module-novel_editor">novel_editor</a>
+</li>
+        <li><a href="processingmenu.html#module-processingmenu">processingmenu</a>
 </li>
       </ul></li>
   </ul></td>
