+++ docs/html/helpmenu.html
@@ -7,7 +7,7 @@
   
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
   
-  <title>Pythonモジュール索引 &mdash; Novel Editor 0.6.0b ドキュメント</title>
+  <title>helpmenu module &mdash; Novel Editor 0.6.0b ドキュメント</title>
   
 
   
@@ -37,13 +37,8 @@
     
     <link rel="index" title="索引" href="genindex.html" />
     <link rel="search" title="検索" href="search.html" />
- 
-
-    <script>
-      DOCUMENTATION_OPTIONS.COLLAPSE_INDEX = true;
-    </script>
-
-
+    <link rel="next" title="mydialog module" href="mydialog.html" />
+    <link rel="prev" title="source" href="modules.html" /> 
 </head>
 
 <body class="wy-body-for-nav">
@@ -89,13 +84,19 @@
             
             
               <p class="caption"><span class="caption-text">Contents:</span></p>
-<ul>
+<ul class="current">
 <li class="toctree-l1"><a class="reference internal" href="use.html">使用方法</a></li>
 <li class="toctree-l1"><a class="reference internal" href="sub/charactor.html">キャラクターファイルについて</a></li>
 <li class="toctree-l1"><a class="reference internal" href="sub/save.html">保存ファイルについて</a></li>
 <li class="toctree-l1"><a class="reference internal" href="sub/exe.html">exe実行ファイルを作成</a></li>
 <li class="toctree-l1"><a class="reference internal" href="sub/coverage.html">カバレッジの計測について</a></li>
-<li class="toctree-l1"><a class="reference internal" href="modules.html">source</a></li>
+<li class="toctree-l1 current"><a class="reference internal" href="modules.html">source</a><ul class="current">
+<li class="toctree-l2 current"><a class="current reference internal" href="#">helpmenu module</a></li>
+<li class="toctree-l2"><a class="reference internal" href="mydialog.html">mydialog module</a></li>
+<li class="toctree-l2"><a class="reference internal" href="novel_editor.html">novel_editor module</a></li>
+<li class="toctree-l2"><a class="reference internal" href="processingmenu.html">processingmenu module</a></li>
+</ul>
+</li>
 </ul>
 
             
@@ -142,10 +143,16 @@
     
       <li><a href="index.html" class="icon icon-home"></a> &raquo;</li>
         
-      <li>Pythonモジュール索引</li>
+          <li><a href="modules.html">source</a> &raquo;</li>
+        
+      <li>helpmenu module</li>
     
     
       <li class="wy-breadcrumbs-aside">
+        
+            
+            <a href="_sources/helpmenu.rst.txt" rel="nofollow"> View page source</a>
+          
         
       </li>
     
@@ -157,56 +164,53 @@
           <div role="main" class="document" itemscope="itemscope" itemtype="http://schema.org/Article">
            <div itemprop="articleBody">
             
+  <div class="section" id="module-helpmenu">
+<span id="helpmenu-module"></span><h1>helpmenu module<a class="headerlink" href="#module-helpmenu" title="このヘッドラインへのパーマリンク">¶</a></h1>
+<dl class="py class">
+<dt id="helpmenu.HelpMenuClass">
+<em class="property">class </em><code class="sig-prename descclassname">helpmenu.</code><code class="sig-name descname">HelpMenuClass</code><a class="reference internal" href="_modules/helpmenu.html#HelpMenuClass"><span class="viewcode-link">[ソース]</span></a><a class="headerlink" href="#helpmenu.HelpMenuClass" title="この定義へのパーマリンク">¶</a></dt>
+<dd><p>ベースクラス: <code class="xref py py-class docutils literal notranslate"><span class="pre">object</span></code></p>
+<dl class="py method">
+<dt id="helpmenu.HelpMenuClass.help">
+<code class="sig-name descname">help</code><span class="sig-paren">(</span><span class="sig-paren">)</span><a class="reference internal" href="_modules/helpmenu.html#HelpMenuClass.help"><span class="viewcode-link">[ソース]</span></a><a class="headerlink" href="#helpmenu.HelpMenuClass.help" title="この定義へのパーマリンク">¶</a></dt>
+<dd><p>helpページを開く</p>
+<p>・ウエブブラウザを使ってREADME.htmlを表示する。</p>
+</dd></dl>
 
-   <h1>Pythonモジュール索引</h1>
+<dl class="py method">
+<dt id="helpmenu.HelpMenuClass.version">
+<code class="sig-name descname">version</code><span class="sig-paren">(</span><em class="sig-param"><span class="n">root</span></em>, <em class="sig-param"><span class="n">image_data</span></em><span class="sig-paren">)</span><a class="reference internal" href="_modules/helpmenu.html#HelpMenuClass.version"><span class="viewcode-link">[ソース]</span></a><a class="headerlink" href="#helpmenu.HelpMenuClass.version" title="この定義へのパーマリンク">¶</a></dt>
+<dd><p>バージョン情報を表示</p>
+<p>・バージョン情報表示ダイアログを表示する。
+×を押すまで消えないようにする。</p>
+<dl class="field-list simple">
+<dt class="field-odd">パラメータ</dt>
+<dd class="field-odd"><ul class="simple">
+<li><p><strong>root</strong> (<em>instance</em>) -- トップレベルのウインドウインスタンス</p></li>
+<li><p><strong>image_data</strong> (<em>str</em>) -- イメージファイル</p></li>
+</ul>
+</dd>
+</dl>
+</dd></dl>
 
-   <div class="modindex-jumpbox">
-   <a href="#cap-h"><strong>h</strong></a> | 
-   <a href="#cap-m"><strong>m</strong></a> | 
-   <a href="#cap-n"><strong>n</strong></a> | 
-   <a href="#cap-p"><strong>p</strong></a>
-   </div>
+</dd></dl>
 
-   <table class="indextable modindextable">
-     <tr class="pcap"><td></td><td>&#160;</td><td></td></tr>
-     <tr class="cap" id="cap-h"><td></td><td>
-       <strong>h</strong></td><td></td></tr>
-     <tr>
-       <td></td>
-       <td>
-       <a href="helpmenu.html#module-helpmenu"><code class="xref">helpmenu</code></a></td><td>
-       <em></em></td></tr>
-     <tr class="pcap"><td></td><td>&#160;</td><td></td></tr>
-     <tr class="cap" id="cap-m"><td></td><td>
-       <strong>m</strong></td><td></td></tr>
-     <tr>
-       <td></td>
-       <td>
-       <a href="mydialog.html#module-mydialog"><code class="xref">mydialog</code></a></td><td>
-       <em></em></td></tr>
-     <tr class="pcap"><td></td><td>&#160;</td><td></td></tr>
-     <tr class="cap" id="cap-n"><td></td><td>
-       <strong>n</strong></td><td></td></tr>
-     <tr>
-       <td></td>
-       <td>
-       <a href="novel_editor.html#module-novel_editor"><code class="xref">novel_editor</code></a></td><td>
-       <em></em></td></tr>
-     <tr class="pcap"><td></td><td>&#160;</td><td></td></tr>
-     <tr class="cap" id="cap-p"><td></td><td>
-       <strong>p</strong></td><td></td></tr>
-     <tr>
-       <td></td>
-       <td>
-       <a href="processingmenu.html#module-processingmenu"><code class="xref">processingmenu</code></a></td><td>
-       <em></em></td></tr>
-   </table>
+</div>
 
 
            </div>
            
           </div>
           <footer>
+  
+    <div class="rst-footer-buttons" role="navigation" aria-label="footer navigation">
+      
+        <a href="mydialog.html" class="btn btn-neutral float-right" title="mydialog module" accesskey="n" rel="next">Next <span class="fa fa-arrow-circle-right"></span></a>
+      
+      
+        <a href="modules.html" class="btn btn-neutral float-left" title="source" accesskey="p" rel="prev"><span class="fa fa-arrow-circle-left"></span> Previous</a>
+      
+    </div>
   
 
   <hr/>
