+++ docs/html/modules.html
@@ -36,7 +36,9 @@
 
     
     <link rel="index" title="索引" href="genindex.html" />
-    <link rel="search" title="検索" href="search.html" /> 
+    <link rel="search" title="検索" href="search.html" />
+    <link rel="next" title="helpmenu module" href="helpmenu.html" />
+    <link rel="prev" title="カバレッジの計測について" href="sub/coverage.html" /> 
 </head>
 
 <body class="wy-body-for-nav">
@@ -82,13 +84,19 @@
             
             
               <p class="caption"><span class="caption-text">Contents:</span></p>
-<ul>
+<ul class="current">
 <li class="toctree-l1"><a class="reference internal" href="use.html">使用方法</a></li>
 <li class="toctree-l1"><a class="reference internal" href="sub/charactor.html">キャラクターファイルについて</a></li>
 <li class="toctree-l1"><a class="reference internal" href="sub/save.html">保存ファイルについて</a></li>
 <li class="toctree-l1"><a class="reference internal" href="sub/exe.html">exe実行ファイルを作成</a></li>
 <li class="toctree-l1"><a class="reference internal" href="sub/coverage.html">カバレッジの計測について</a></li>
-<li class="toctree-l1"><a class="reference internal" href="novel_editor.html">Novel Editor モジュール</a></li>
+<li class="toctree-l1 current"><a class="current reference internal" href="#">source</a><ul>
+<li class="toctree-l2"><a class="reference internal" href="helpmenu.html">helpmenu module</a></li>
+<li class="toctree-l2"><a class="reference internal" href="mydialog.html">mydialog module</a></li>
+<li class="toctree-l2"><a class="reference internal" href="novel_editor.html">novel_editor module</a></li>
+<li class="toctree-l2"><a class="reference internal" href="processingmenu.html">processingmenu module</a></li>
+</ul>
+</li>
 </ul>
 
             
@@ -158,7 +166,10 @@
 <h1>source<a class="headerlink" href="#source" title="このヘッドラインへのパーマリンク">¶</a></h1>
 <div class="toctree-wrapper compound">
 <ul>
-<li class="toctree-l1"><a class="reference internal" href="novel_editor.html">Novel Editor モジュール</a></li>
+<li class="toctree-l1"><a class="reference internal" href="helpmenu.html">helpmenu module</a></li>
+<li class="toctree-l1"><a class="reference internal" href="mydialog.html">mydialog module</a></li>
+<li class="toctree-l1"><a class="reference internal" href="novel_editor.html">novel_editor module</a></li>
+<li class="toctree-l1"><a class="reference internal" href="processingmenu.html">processingmenu module</a></li>
 </ul>
 </div>
 </div>
@@ -168,6 +179,15 @@
            
           </div>
           <footer>
+  
+    <div class="rst-footer-buttons" role="navigation" aria-label="footer navigation">
+      
+        <a href="helpmenu.html" class="btn btn-neutral float-right" title="helpmenu module" accesskey="n" rel="next">Next <span class="fa fa-arrow-circle-right"></span></a>
+      
+      
+        <a href="sub/coverage.html" class="btn btn-neutral float-left" title="カバレッジの計測について" accesskey="p" rel="prev"><span class="fa fa-arrow-circle-left"></span> Previous</a>
+      
+    </div>
   
 
   <hr/>
