+++ docs/html/processingmenu.html
@@ -7,7 +7,7 @@
   
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
   
-  <title>Pythonモジュール索引 &mdash; Novel Editor 0.6.0b ドキュメント</title>
+  <title>processingmenu module &mdash; Novel Editor 0.6.0b ドキュメント</title>
   
 
   
@@ -37,13 +37,7 @@
     
     <link rel="index" title="索引" href="genindex.html" />
     <link rel="search" title="検索" href="search.html" />
- 
-
-    <script>
-      DOCUMENTATION_OPTIONS.COLLAPSE_INDEX = true;
-    </script>
-
-
+    <link rel="prev" title="novel_editor module" href="novel_editor.html" /> 
 </head>
 
 <body class="wy-body-for-nav">
@@ -89,13 +83,19 @@
             
             
               <p class="caption"><span class="caption-text">Contents:</span></p>
-<ul>
+<ul class="current">
 <li class="toctree-l1"><a class="reference internal" href="use.html">使用方法</a></li>
 <li class="toctree-l1"><a class="reference internal" href="sub/charactor.html">キャラクターファイルについて</a></li>
 <li class="toctree-l1"><a class="reference internal" href="sub/save.html">保存ファイルについて</a></li>
 <li class="toctree-l1"><a class="reference internal" href="sub/exe.html">exe実行ファイルを作成</a></li>
 <li class="toctree-l1"><a class="reference internal" href="sub/coverage.html">カバレッジの計測について</a></li>
-<li class="toctree-l1"><a class="reference internal" href="modules.html">source</a></li>
+<li class="toctree-l1 current"><a class="reference internal" href="modules.html">source</a><ul class="current">
+<li class="toctree-l2"><a class="reference internal" href="helpmenu.html">helpmenu module</a></li>
+<li class="toctree-l2"><a class="reference internal" href="mydialog.html">mydialog module</a></li>
+<li class="toctree-l2"><a class="reference internal" href="novel_editor.html">novel_editor module</a></li>
+<li class="toctree-l2 current"><a class="current reference internal" href="#">processingmenu module</a></li>
+</ul>
+</li>
 </ul>
 
             
@@ -142,10 +142,16 @@
     
       <li><a href="index.html" class="icon icon-home"></a> &raquo;</li>
         
-      <li>Pythonモジュール索引</li>
+          <li><a href="modules.html">source</a> &raquo;</li>
+        
+      <li>processingmenu module</li>
     
     
       <li class="wy-breadcrumbs-aside">
+        
+            
+            <a href="_sources/processingmenu.rst.txt" rel="nofollow"> View page source</a>
+          
         
       </li>
     
@@ -157,56 +163,48 @@
           <div role="main" class="document" itemscope="itemscope" itemtype="http://schema.org/Article">
            <div itemprop="articleBody">
             
+  <div class="section" id="module-processingmenu">
+<span id="processingmenu-module"></span><h1>processingmenu module<a class="headerlink" href="#module-processingmenu" title="このヘッドラインへのパーマリンク">¶</a></h1>
+<dl class="py class">
+<dt id="processingmenu.ProcessingMenuClass">
+<em class="property">class </em><code class="sig-prename descclassname">processingmenu.</code><code class="sig-name descname">ProcessingMenuClass</code><span class="sig-paren">(</span><em class="sig-param"><span class="n">app</span></em><span class="sig-paren">)</span><a class="reference internal" href="_modules/processingmenu.html#ProcessingMenuClass"><span class="viewcode-link">[ソース]</span></a><a class="headerlink" href="#processingmenu.ProcessingMenuClass" title="この定義へのパーマリンク">¶</a></dt>
+<dd><p>ベースクラス: <code class="xref py py-class docutils literal notranslate"><span class="pre">object</span></code></p>
+<dl class="py method">
+<dt id="processingmenu.ProcessingMenuClass.find_wikipedia">
+<code class="sig-name descname">find_wikipedia</code><span class="sig-paren">(</span><em class="sig-param"><span class="n">wiki_wiki</span></em><span class="sig-paren">)</span><a class="reference internal" href="_modules/processingmenu.html#ProcessingMenuClass.find_wikipedia"><span class="viewcode-link">[ソース]</span></a><a class="headerlink" href="#processingmenu.ProcessingMenuClass.find_wikipedia" title="この定義へのパーマリンク">¶</a></dt>
+<dd><p>意味を検索</p>
+<p>・Wikipedia-APIライブラリを使ってWikipediaから選択文字の意味を
+検索する。</p>
+<dl class="field-list simple">
+<dt class="field-odd">パラメータ</dt>
+<dd class="field-odd"><p><strong>wiki_wiki</strong> (<em>instance</em>) -- Wikipedia-APIのインスタンス</p>
+</dd>
+</dl>
+</dd></dl>
 
-   <h1>Pythonモジュール索引</h1>
+<dl class="py method">
+<dt id="processingmenu.ProcessingMenuClass.open_becoming_novelist_page">
+<code class="sig-name descname">open_becoming_novelist_page</code><span class="sig-paren">(</span><span class="sig-paren">)</span><a class="reference internal" href="_modules/processingmenu.html#ProcessingMenuClass.open_becoming_novelist_page"><span class="viewcode-link">[ソース]</span></a><a class="headerlink" href="#processingmenu.ProcessingMenuClass.open_becoming_novelist_page" title="この定義へのパーマリンク">¶</a></dt>
+<dd><p>小説家になろうのユーザーページを開く</p>
+<p>・インターネットブラウザで小説家になろうのユーザーページを開く。</p>
+</dd></dl>
 
-   <div class="modindex-jumpbox">
-   <a href="#cap-h"><strong>h</strong></a> | 
-   <a href="#cap-m"><strong>m</strong></a> | 
-   <a href="#cap-n"><strong>n</strong></a> | 
-   <a href="#cap-p"><strong>p</strong></a>
-   </div>
+</dd></dl>
 
-   <table class="indextable modindextable">
-     <tr class="pcap"><td></td><td>&#160;</td><td></td></tr>
-     <tr class="cap" id="cap-h"><td></td><td>
-       <strong>h</strong></td><td></td></tr>
-     <tr>
-       <td></td>
-       <td>
-       <a href="helpmenu.html#module-helpmenu"><code class="xref">helpmenu</code></a></td><td>
-       <em></em></td></tr>
-     <tr class="pcap"><td></td><td>&#160;</td><td></td></tr>
-     <tr class="cap" id="cap-m"><td></td><td>
-       <strong>m</strong></td><td></td></tr>
-     <tr>
-       <td></td>
-       <td>
-       <a href="mydialog.html#module-mydialog"><code class="xref">mydialog</code></a></td><td>
-       <em></em></td></tr>
-     <tr class="pcap"><td></td><td>&#160;</td><td></td></tr>
-     <tr class="cap" id="cap-n"><td></td><td>
-       <strong>n</strong></td><td></td></tr>
-     <tr>
-       <td></td>
-       <td>
-       <a href="novel_editor.html#module-novel_editor"><code class="xref">novel_editor</code></a></td><td>
-       <em></em></td></tr>
-     <tr class="pcap"><td></td><td>&#160;</td><td></td></tr>
-     <tr class="cap" id="cap-p"><td></td><td>
-       <strong>p</strong></td><td></td></tr>
-     <tr>
-       <td></td>
-       <td>
-       <a href="processingmenu.html#module-processingmenu"><code class="xref">processingmenu</code></a></td><td>
-       <em></em></td></tr>
-   </table>
+</div>
 
 
            </div>
            
           </div>
           <footer>
+  
+    <div class="rst-footer-buttons" role="navigation" aria-label="footer navigation">
+      
+      
+        <a href="novel_editor.html" class="btn btn-neutral float-left" title="novel_editor module" accesskey="p" rel="prev"><span class="fa fa-arrow-circle-left"></span> Previous</a>
+      
+    </div>
   
 
   <hr/>
