+++ docs/html/search.html
@@ -90,7 +90,7 @@
 <li class="toctree-l1"><a class="reference internal" href="sub/save.html">保存ファイルについて</a></li>
 <li class="toctree-l1"><a class="reference internal" href="sub/exe.html">exe実行ファイルを作成</a></li>
 <li class="toctree-l1"><a class="reference internal" href="sub/coverage.html">カバレッジの計測について</a></li>
-<li class="toctree-l1"><a class="reference internal" href="novel_editor.html">Novel Editor モジュール</a></li>
+<li class="toctree-l1"><a class="reference internal" href="modules.html">source</a></li>
 </ul>
