+++ docs/html/searchindex.js
@@ -1,1 +1,1 @@
-Search.setIndex({docnames:["index","modules","novel_editor","sub/charactor","sub/coverage","sub/exe","sub/save","use"],envversion:{"sphinx.domains.c":2,"sphinx.domains.changeset":1,"sphinx.domains.citation":1,"sphinx.domains.cpp":3,"sphinx.domains.index":1,"sphinx.domains.javascript":2,"sphinx.domains.math":2,"sphinx.domains.python":2,"sphinx.domains.rst":2,"sphinx.domains.std":1,"sphinx.ext.viewcode":1,sphinx:56},filenames:["index.rst","modules.rst","novel_editor.rst","sub\\charactor.md","sub\\coverage.md","sub\\exe.md","sub\\save.md","use.md"],objects:{"":{novel_editor:[2,0,0,"-"]},"novel_editor.LineFrame":{all_highlight:[2,2,1,""],auto_complete:[2,2,1,""],btn_click:[2,2,1,""],change_setting:[2,2,1,""],clear_btn_click:[2,2,1,""],copy:[2,2,1,""],create_event:[2,2,1,""],create_event_character:[2,2,1,""],create_event_image:[2,2,1,""],create_event_text:[2,2,1,""],create_tags:[2,2,1,""],create_widgets:[2,2,1,""],cut:[2,2,1,""],externalLoop:[2,2,1,""],find_dialog:[2,2,1,""],find_dictionaly:[2,2,1,""],font_dialog:[2,2,1,""],font_size_Change:[2,2,1,""],frame:[2,2,1,""],frame_character:[2,2,1,""],frame_image:[2,2,1,""],get_current_insert_word:[2,2,1,""],get_keywords:[2,2,1,""],initialize:[2,2,1,""],is_hiragana:[2,2,1,""],line_highlight:[2,2,1,""],message_window:[2,2,1,""],moji_count:[2,2,1,""],mouse_image_scroll:[2,2,1,""],mouse_y_scroll:[2,2,1,""],new_file:[2,2,1,""],new_open:[2,2,1,""],on_double_click:[2,2,1,""],on_double_click_yahoo:[2,2,1,""],on_name_click:[2,2,1,""],open_file:[2,2,1,""],open_file_save:[2,2,1,""],open_help:[2,2,1,""],open_url:[2,2,1,""],open_version:[2,2,1,""],overwrite_save_file:[2,2,1,""],paste:[2,2,1,""],path_read_image:[2,2,1,""],path_read_text:[2,2,1,""],print_gif:[2,2,1,""],push_keys:[2,2,1,""],pyttsx3_onend:[2,2,1,""],pyttsx3_onreadend:[2,2,1,""],pyttsx3_onword:[2,2,1,""],read_text:[2,2,1,""],redo:[2,2,1,""],remove_list:[2,2,1,""],replacement:[2,2,1,""],replacement_dialog:[2,2,1,""],replacement_dialog_on_closing:[2,2,1,""],resize_gif:[2,2,1,""],ruby:[2,2,1,""],save_charactor_file:[2,2,1,""],save_file:[2,2,1,""],search:[2,2,1,""],search_forward:[2,2,1,""],search_next:[2,2,1,""],selection:[2,2,1,""],tab:[2,2,1,""],tree_get_loop:[2,2,1,""],undo:[2,2,1,""],update_line_numbers:[2,2,1,""],yahoo:[2,2,1,""],yahoocall:[2,2,1,""],yahooresult:[2,2,1,""]},"novel_editor.Mydialog":{sub_name_ok:[2,2,1,""]},"novel_editor.Mydialog.self":{sub_name_win:[2,3,1,""],txt:[2,3,1,""],txt_name:[2,3,1,""]},novel_editor:{CustomText:[2,1,1,""],LineFrame:[2,1,1,""],Mydialog:[2,1,1,""],on_closing:[2,4,1,""]}},objnames:{"0":["py","module","Python \u30e2\u30b8\u30e5\u30fc\u30eb"],"1":["py","class","Python \u30af\u30e9\u30b9"],"2":["py","method","Python \u30e1\u30bd\u30c3\u30c9"],"3":["py","attribute","Python \u306e\u5c5e\u6027"],"4":["py","function","Python \u306e\u95a2\u6570"]},objtypes:{"0":"py:module","1":"py:class","2":"py:method","3":"py:attribute","4":"py:function"},terms:{"(%":2,"(http":7,"(true":2,"(wikipedia":7,")<":3,")\u3001":6,"--":2,"-api":[2,7],".bat":5,".event":2,".frame":2,".html":2,".ned":6,".py":4,".py\u30d5\u30a1\u30a4\u30eb":7,".text":2,".ttk":2,"/novel":4,"/py":5,"2exe":5,"</":3,"<<":2,"==":[5,7],"=\"":3,">>":2,">\n<":3,"?xml":3,"\"?":3,"\u3001.":6,"\u3001<":2,"\u3002(":[6,7],"\u3002)":[6,7],"\u3042\u308a":6,"\u3042\u308b":6,"\u3042\u308c":2,"\u3044\u308b":[2,3],"\u304a\u3053\u306a\u3046":2,"\u304a\u9858\u3044":7,"\u304b\u304b\u308b":2,"\u304b\u3068":7,"\u304b\u3089":2,"\u304c\u308f\u304b\u308b":2,"\u304f\u3060":[5,7],"\u3053\u3068":[2,7],"\u3053\u308c":7,"\u3055\u3044":[5,7],"\u3059\u3079\u3066":2,"\u3059\u308b":[2,4,5,6],"\u3059\u308c":[2,7],"\u305d\u3053\u306b":6,"\u305d\u306e":[2,3,4],"\u305d\u308c":7,"\u305d\u308c\u305e\u308c":[2,6],"\u305f\u3044":2,"\u305f\u3081":[2,4,5],"\u305f\u3089":2,"\u3060\u3051":[2,6],"\u3064\u304e":4,"\u3064\u3051":[2,6],"\u3066\u308b":2,"\u3067\u304d":[2,6,7],"\u3067\u304d\u308b":7,"\u3067\u304d\u308c":2,"\u3067\u3059":[5,6,7],"\u3068\u304d":[2,6],"\u3069\u3046":2,"\u306a\u3044":[2,7],"\u306a\u304b":2,"\u306a\u3051\u308c":2,"\u306a\u306a\u3089":2,"\u306a\u3089":2,"\u306a\u308d":2,"\u306b\u3064\u3044\u3066":0,"\u306b\u3088\u3063\u3066":6,"\u306e\u3067":[5,6,7],"\u306e\u307f":2,"\u306f\u305a":7,"\u3072\u3089":2,"\u3075\u308b":2,"\u307e\u3057":7,"\u307e\u3059":[4,5,6,7],"\u307e\u305b":7,"\u307e\u3067":2,"\u3082\u306e":2,"\u3084\u3081":2,"\u3088\u3046":2,"\u3089\u308c":2,"\u308c\u308b":2,"\u30a2\u30a4\u30c6\u30e0":2,"\u30a4\u30d9\u30f3\u30c8":[2,6],"\u30a4\u30e1\u30fc\u30b8":[2,6],"\u30a4\u30e1\u30fc\u30b8\u30a4\u30d9\u30f3\u30c8":2,"\u30a4\u30e1\u30fc\u30b8\u30a4\u30f3\u30b9\u30bf\u30f3\u30b9":2,"\u30a4\u30e1\u30fc\u30b8\u30ad\u30e3\u30f3\u30d0\u30b9":2,"\u30a4\u30e1\u30fc\u30b8\u30d5\u30a1\u30a4\u30eb":2,"\u30a4\u30e1\u30fc\u30b8\u30d5\u30ec\u30fc\u30e0":2,"\u30a4\u30f3\u30b9\u30bf\u30f3\u30b9":2,"\u30a4\u30f3\u30b9\u30c8\u30fc\u30eb":[4,5,7],"\u30a4\u30f3\u30d7\u30c3\u30c8\u30dc\u30c3\u30af\u30b9":2,"\u30a4\u30f3\u30d7\u30c3\u30c8\u30dc\u30c3\u30af\u30b9\u30a4\u30f3\u30b9\u30bf\u30f3\u30b9":2,"\u30a6\u30a3\u30b8\u30a7\u30c3\u30c8":2,"\u30a6\u30a4\u30f3\u30c9\u30a6":2,"\u30a6\u30a7\u30b8\u30c3\u30c8":2,"\u30a8\u30f3\u30b8\u30f3":7,"\u30ab\u30a6\u30f3\u30c8":2,"\u30ab\u30d0\u30ec\u30c3\u30b8":0,"\u30ab\u30d0\u30ec\u30c3\u30b8\u30ec\u30dd\u30fc\u30c8":4,"\u30ad\u30e3\u30e9\u30af\u30bf\u30fc":[2,6],"\u30ad\u30e3\u30e9\u30af\u30bf\u30fc\u30d5\u30a1\u30a4\u30eb":[0,2],"\u30ad\u30e3\u30f3\u30d0\u30b9":2,"\u30ad\u30fc":2,"\u30ad\u30fc\u30ef\u30fc\u30c9":2,"\u30af\u30e9\u30b9":2,"\u30af\u30ea\u30c3\u30af":2,"\u30b3\u30de\u30f3\u30c9":4,"\u30b3\u30fc\u30c9":[4,5,7],"\u30b5\u30a4\u30ba":2,"\u30b9\u30af\u30ed\u30fc\u30eb":2,"\u30b9\u30af\u30ed\u30fc\u30eb\u30d0\u30fc":2,"\u30bb\u30fc\u30d6\u30d5\u30a1\u30a4\u30eb":3,"\u30bd\u30d5\u30c8":6,"\u30bd\u30fc\u30b9":2,"\u30bf\u30a4\u30c8\u30eb":2,"\u30bf\u30b0":2,"\u30bf\u30d6":2,"\u30bf\u30fc\u30df\u30ca\u30eb\u30a8\u30c7\u30a3\u30bf":[5,7],"\u30c0\u30a4\u30a2\u30ed\u30b0":2,"\u30c0\u30a4\u30a2\u30ed\u30b0\u30a6\u30a4\u30f3\u30c9\u30a6\u30a4\u30f3\u30b9\u30bf\u30f3\u30b9":2,"\u30c0\u30a4\u30a2\u30ed\u30b0\u30dc\u30bf\u30f3\u30af\u30ea\u30c3\u30af":2,"\u30c0\u30d6\u30eb\u30af\u30ea\u30c3\u30af":2,"\u30c4\u30ea\u30fc\u30d3\u30e5\u30fc":2,"\u30c4\u30ea\u30fc\u30d3\u30e5\u30fc\u30a4\u30d9\u30f3\u30c8":2,"\u30c6\u30ad\u30b9\u30c8":[2,6],"\u30c6\u30ad\u30b9\u30c8\u30a4\u30d9\u30f3\u30c8":2,"\u30c6\u30ad\u30b9\u30c8\u30d5\u30a1\u30a4\u30eb":[2,6],"\u30c6\u30ad\u30b9\u30c8\u30d5\u30a9\u30eb\u30c0":6,"\u30c6\u30ad\u30b9\u30c8\u30dc\u30c3\u30af\u30b9":2,"\u30c7\u30fc\u30bf":[2,6],"\u30cf\u30a4\u30e9\u30a4\u30c8":2,"\u30d0\u30c3\u30c1\u30d5\u30a1\u30a4\u30eb":5,"\u30d0\u30fc\u30b8\u30e7\u30f3":[2,7],"\u30d1\u30b9":2,"\u30d1\u30e9\u30e1\u30fc\u30bf":2,"\u30d5\u30a1\u30a4\u30eb":[0,2,3],"\u30d5\u30a9\u30eb\u30c0":[3,4,5,6,7],"\u30d5\u30a9\u30f3\u30c8":2,"\u30d5\u30a9\u30f3\u30c8\u30b5\u30a4\u30ba\u30c0\u30a4\u30a2\u30ed\u30b0":2,"\u30d5\u30a9\u30fc\u30ab\u30b9":2,"\u30d5\u30ec\u30fc\u30e0":2,"\u30d9\u30fc\u30b9\u30af\u30e9\u30b9":2,"\u30da\u30fc\u30b8":2,"\u30dc\u30bf\u30f3":2,"\u30dc\u30c3\u30af\u30b9":2,"\u30de\u30a6\u30b9\u30db\u30a4\u30fc\u30eb":2,"\u30e1\u30a4\u30f3":2,"\u30e1\u30a4\u30f3\u30bd\u30fc\u30b9\u30d5\u30a1\u30a4\u30eb":7,"\u30e1\u30a4\u30f3\u30d5\u30ec\u30fc\u30e0":2,"\u30e2\u30b8\u30e5\u30fc\u30eb":[0,1],"\u30e6\u30fc\u30b6\u30fc\u30da\u30fc\u30b8":2,"\u30e9\u30a4\u30d6\u30e9\u30ea":[2,7],"\u30ea\u30b5\u30a4\u30ba":2,"\u30ea\u30b9\u30c8":2,"\u30eb\u30d3":2,"\u30fb\u305d\u308c\u305e\u308c":2,"\u30fb\u30a4\u30e1\u30fc\u30b8\u30ad\u30e3\u30f3\u30d0\u30b9":2,"\u30fb\u30a4\u30e1\u30fc\u30b8\u30d5\u30a1\u30a4\u30eb":2,"\u30fb\u30a4\u30f3\u30bf\u30fc\u30cd\u30c3\u30c8\u30d6\u30e9\u30a6\u30b6":2,"\u30fb\u30a6\u30a8\u30d6\u30d6\u30e9\u30a6\u30b6":2,"\u30fb\u30ad\u30e3\u30e9\u30af\u30bf\u30fc":2,"\u30fb\u30b5\u30a4\u30ba":2,"\u30fb\u30bd\u30d5\u30c8":2,"\u30fb\u30bf\u30d6\u30ad\u30fc":2,"\u30fb\u30c4\u30ea\u30fc\u30d3\u30e5\u30fc":2,"\u30fb\u30c6\u30ad\u30b9\u30c8":2,"\u30fb\u30c6\u30ad\u30b9\u30c8\u30dc\u30c3\u30af\u30b9":2,"\u30fb\u30d0\u30fc\u30b8\u30e7\u30f3":2,"\u30fb\u30d1\u30b9":2,"\u30fb\u30d5\u30a1\u30a4\u30eb":2,"\u30fb\u30d5\u30a9\u30f3\u30c8\u30b5\u30a4\u30ba\u30c0\u30a4\u30a2\u30ed\u30b0":2,"\u30fb\u30e1\u30a4\u30f3":2,"\u30fb\u30e1\u30a4\u30f3\u30a6\u30a4\u30f3\u30c9\u30a6":2,"\u30fb\u30ea\u30b9\u30c8\u30dc\u30c3\u30af\u30b9":2,"\u30fb\u4e0e\u3048":2,"\u30fb\u4f3c\u9854\u7d75":2,"\u30fb\u4f55":2,"\u30fb\u4fdd\u5b58":2,"\u30fb\u5165\u529b\u7b49":2,"\u30fb\u5909\u6570":2,"\u30fb\u5b50\u30a2\u30a4\u30c6\u30e0":2,"\u30fb\u6587\u5b57\u6570":2,"\u30fb\u6587\u7ae0\u8aad":2,"\u30fb\u6607\u9806":2,"\u30fb\u6821\u6b63":2,"\u30fb\u691c\u7d22":2,"\u30fb\u6d88\u53bb":2,"\u30fb\u73fe\u5728":2,"\u30fb\u7f6e\u63db":2,"\u30fb\u81ea\u4f5c\u30c0\u30a4\u30a2\u30ed\u30b0":2,"\u30fb\u884c\u756a\u53f7":2,"\u30fb\u88dc\u5b8c":2,"\u30fb\u958b\u3044":2,"\u30fb\u958b\u304f":2,"\u30fbcopy":2,"\u30fbcut":2,"\u30fbpaste":2,"\u30fbpyttsx":2,"\u30fbredo":2,"\u30fbtext":2,"\u30fbuedo":2,"\u30fbwikipedia":2,"\u30fbyahoo":2,"\u4e0a\u3052":2,"\u4e0a\u66f8":2,"\u4e0a\u66f8\u304d":2,"\u4e0a\u8a18":7,"\u4e0b\u8a18":[5,7],"\u4e0d\u660e":2,"\u4e21\u65b9":2,"\u4e2d\u65ad":2,"\u4e2d\u6b62":2,"\u4ed8\u3051":2,"\u4ed8\u3051\u308b":2,"\u4ee5\u5916":7,"\u4f3c\u9854":[2,6],"\u4f3c\u9854\u7d75":2,"\u4f46\u3057":[6,7],"\u4f4d\u7f6e":2,"\u4f5c\u6210":[0,2],"\u4f7f\u3044":5,"\u4f7f\u3046":6,"\u4f7f\u3063":[2,7],"\u4f7f\u7528":[0,4,5],"\u4fdd\u5b58":[0,2,3],"\u5019\u88dc":2,"\u5165\u3063":[3,7],"\u5165\u529b":[2,4],"\u51e6\u7406":[2,7],"\u51fa\u3057":2,"\u51fa\u3059":2,"\u51fa\u529b":4,"\u521d\u3081":2,"\u521d\u671f":2,"\u521d\u671f\u5316":2,"\u5224\u65ad":2,"\u5236\u4f5c":2,"\u524a\u9664":2,"\u5316\u3051":6,"\u5316\u53ca":2,"\u5358\u8a9e":2,"\u53d6\u308c":7,"\u53d6\u5f97":2,"\u53f3\u5074":2,"\u5408\u3048":7,"\u5408\u6210":7,"\u540c\u3058":[2,6],"\u540d\u524d":[2,6],"\u547c\u3073\u51fa\u3057":2,"\u547c\u3073\u51fa\u3059":2,"\u547c\u3073\u540d":[2,3],"\u56de\u3057":2,"\u5834\u6240":[2,6],"\u5909\u3048\u308b":2,"\u5909\u63db":[5,7],"\u5909\u66f4":[2,6],"\u5927\u9805\u76ee":6,"\u59cb\u3081\u308b":2,"\u5b58\u5728":2,"\u5b9f\u4f53":6,"\u5b9f\u884c":[0,4,7],"\u5c0b\u306d":2,"\u5c0f\u8aac":6,"\u5c0f\u8aac\u5bb6":2,"\u5c0f\u9805\u76ee":6,"\u5f35\u308a":2,"\u5f62\u5f0f":[2,3,6],"\u5f62\u614b":[2,7],"\u5f8c\u4ee5":4,"\u5f97\u308b":2,"\u601d\u3044":7,"\u6027\u5225":3,"\u6050\u308c":6,"\u60c5\u5831":2,"\u610f\u5473":2,"\u623b\u3059":[2,6],"\u623b\u308a\u5024":2,"\u62bc\u3055":2,"\u62bc\u3057":2,"\u62bc\u3059":2,"\u62bc\u4e0b":2,"\u62e1\u5927":[2,6],"\u62e1\u5927\u7387":2,"\u62e1\u5f35":2,"\u62e1\u5f35\u5b50":6,"\u632f\u308b":2,"\u633f\u5165":2,"\u63cf\u753b":2,"\u652f\u63f4":2,"\u6570\u5b57":6,"\u6587\u5b57":[2,6,7],"\u6587\u5b57\u5217":2,"\u6587\u5b57\u6570":2,"\u6587\u7ae0":[2,7],"\u6587\u7ae0\u8aad":2,"\u65b0\u898f":2,"\u65b9\u6cd5":[0,4],"\u6607\u9806":2,"\u66f8\u304b":2,"\u6700\u4e2d":2,"\u6700\u521d":2,"\u672c\u540d":3,"\u6821\u6b63":2,"\u691c\u7d22":[2,7],"\u6a19\u6e96":7,"\u6b62\u3081\u308b":2,"\u6d88\u3048":2,"\u6d88\u53bb":2,"\u6e96\u5099":2,"\u72b6\u614b":2,"\u73fe\u5728":2,"\u73fe\u5728\u884c":2,"\u74b0\u5883":7,"\u753b\u50cf":[2,6,7],"\u753b\u9762":2,"\u7565\u6b74":[2,3],"\u76f8\u5bfe":2,"\u78ba\u8a8d":[2,7],"\u79fb\u52d5":2,"\u7a7a\u6b04":2,"\u7b87\u6240":2,"\u7d20\u89e3":[2,7],"\u7d22\u5f15":0,"\u7d42\u3048":2,"\u7d42\u308f\u3063":2,"\u7d42\u4e86":2,"\u7d50\u679c":2,"\u7d9a\u3051\u308b":2,"\u7e2e\u5c0f":2,"\u7e70\u308a\u8fd4\u3057":2,"\u7f6e\u304f":6,"\u7f6e\u63db":2,"\u8077\u7a2e":6,"\u81ea\u5df1":7,"\u8272\u4ed8\u304d":2,"\u884c\u3046":2,"\u884c\u6570":2,"\u884c\u756a":2,"\u884c\u756a\u53f7":2,"\u8868\u793a":[2,6],"\u88dc\u5b8c":2,"\u89e3\u51cd":6,"\u8a08\u6e2c":0,"\u8a18\u5165":[5,7],"\u8a2d\u5b9a":2,"\u8a72\u5f53":2,"\u8a73\u7d30":4,"\u8a95\u751f":[2,3],"\u8aad\u307f":2,"\u8aad\u307f\u4e0a\u3052":[2,7],"\u8aad\u307f\u4e0a\u3052\u308b":2,"\u8aad\u307f\u8fbc\u3093":2,"\u8cac\u4efb":7,"\u8cbc\u308a\u4ed8\u3051\u308b":2,"\u8d77\u52d5":[2,5,7],"\u8ffd\u52a0":2,"\u9055\u3046":2,"\u9078\u629e":2,"\u914d\u7f6e":2,"\u9577\u3044":2,"\u9589\u3058":2,"\u9589\u3058\u308b":2,"\u958b\u3044":2,"\u958b\u304f":2,"\u958b\u767a":7,"\u95a2\u4fc2":2,"\u95a2\u9023":2,"\u97f3\u58f0":7,"\u9805\u76ee":2,"case":2,"char":2,"class":2,"false":2,"int":2,"new":2,"true":2,_btn:2,_change:2,_character:2,_charactor:2,_click:2,_closing:2,_complete:2,_count:2,_current:2,_dialog:2,_dictionaly:2,_double:2,_editor:[2,4],_event:2,_file:2,_forward:2,_get:2,_gif:2,_help:2,_highlight:2,_hiragana:2,_image:2,_insert:2,_keys:2,_keywords:2,_line:2,_list:2,_loop:2,_name:2,_next:2,_numbers:2,_ok:2,_on:2,_open:2,_path:2,_read:2,_save:2,_scroll:2,_setting:2,_size:2,_tags:2,_text:2,_url:2,_version:2,_widgets:2,_win:2,_window:2,_word:2,_y:2,_yahoo:2,all:2,appid:2,auto:2,birthday:3,bit:7,body:3,bool:2,btn:2,button:2,call:3,change:2,character:[3,6],clear:2,client:2,completed:2,contents:0,copy:2,coverage:4,coveragepy:4,create:2,ctrl:2,customtext:2,cut:2,data:[3,6],dist:5,editor:[1,7],event:[2,6],ex:2,exe:0,externalloop:2,find:2,font:2,frame:2,get:2,gif:[2,6],help:2,html:[2,4],htmlcov:4,id:2,im:2,image:[2,6],index:2,initialize:2,install:[4,5,7],instance:2,is:2,jaconv:7,janome:[2,7],kwargs:2,length:2,line:2,lineframe:2,location:2,master:2,message:2,moji:2,mouse:2,mydialog:2,name:[2,3],nobel:6,none:2,novel:[1,7],object:2,occupation:6,on:2,onend:2,onreadend:2,onword:2,open:2,overwrite:2,paste:2,path:2,pillow:7,pip:[4,5,7],print:2,pt:2,push:2,px:2,py:5,pyinstaller:5,python:[4,7],pyttsx:[2,7],read:2,readme:2,redo:2,remove:2,replacement:2,requests:7,resize:2,ruby:2,run:4,save:2,scale:2,scroll:2,search:2,selection:2,self:2,sentence:2,sex:3,source:[4,7],space:6,str:2,sub:2,tab:2,text:2,title:2,tkinter:2,tree:2,txt:2,type:2,ubuntu:7,undo:2,update:2,version:[3,7],wikipedia:[2,7],windows:[5,7],xml:2,yahoo:2,yahoocall:2,yahooresult:2,zip:6},titles:["Welcome to Novel Editor's documentation!","source","Novel Editor \u30e2\u30b8\u30e5\u30fc\u30eb","\u30ad\u30e3\u30e9\u30af\u30bf\u30fc\u30d5\u30a1\u30a4\u30eb\u306b\u3064\u3044\u3066","\u30ab\u30d0\u30ec\u30c3\u30b8\u306e\u8a08\u6e2c\u306b\u3064\u3044\u3066","exe\u5b9f\u884c\u30d5\u30a1\u30a4\u30eb\u3092\u4f5c\u6210","\u4fdd\u5b58\u30d5\u30a1\u30a4\u30eb\u306b\u3064\u3044\u3066","\u4f7f\u7528\u65b9\u6cd5"],titleterms:{"'s":0,"\u306b\u3064\u3044\u3066":[3,4,6],"\u30ab\u30d0\u30ec\u30c3\u30b8":4,"\u30ad\u30e3\u30e9\u30af\u30bf\u30fc\u30d5\u30a1\u30a4\u30eb":3,"\u30d5\u30a1\u30a4\u30eb":[5,6],"\u30e2\u30b8\u30e5\u30fc\u30eb":2,"\u4f5c\u6210":5,"\u4f7f\u7528":7,"\u4fdd\u5b58":6,"\u5b9f\u884c":5,"\u65b9\u6cd5":7,"\u8a08\u6e2c":4,_editor:[],and:0,documentation:0,editor:[0,2],exe:5,indices:0,novel:[0,2],source:1,tables:0,to:0,welcome:0}})
+Search.setIndex({docnames:["helpmenu","index","modules","mydialog","novel_editor","processingmenu","sub/charactor","sub/coverage","sub/exe","sub/save","use"],envversion:{"sphinx.domains.c":2,"sphinx.domains.changeset":1,"sphinx.domains.citation":1,"sphinx.domains.cpp":3,"sphinx.domains.index":1,"sphinx.domains.javascript":2,"sphinx.domains.math":2,"sphinx.domains.python":2,"sphinx.domains.rst":2,"sphinx.domains.std":1,"sphinx.ext.viewcode":1,sphinx:56},filenames:["helpmenu.rst","index.rst","modules.rst","mydialog.rst","novel_editor.rst","processingmenu.rst","sub\\charactor.md","sub\\coverage.md","sub\\exe.md","sub\\save.md","use.md"],objects:{"":{helpmenu:[0,0,0,"-"],mydialog:[3,0,0,"-"],novel_editor:[4,0,0,"-"],processingmenu:[5,0,0,"-"]},"helpmenu.HelpMenuClass":{help:[0,2,1,""],version:[0,2,1,""]},"mydialog.Mydialog":{sub_name_ok:[3,2,1,""]},"mydialog.Mydialog.self":{sub_name_win:[3,3,1,""],txt:[3,3,1,""],txt_name:[3,3,1,""]},"novel_editor.LineFrame":{all_highlight:[4,2,1,""],auto_complete:[4,2,1,""],btn_click:[4,2,1,""],change_setting:[4,2,1,""],clear_btn_click:[4,2,1,""],copy:[4,2,1,""],create_event:[4,2,1,""],create_event_character:[4,2,1,""],create_event_image:[4,2,1,""],create_event_text:[4,2,1,""],create_tags:[4,2,1,""],create_widgets:[4,2,1,""],cut:[4,2,1,""],externalLoop:[4,2,1,""],find_dialog:[4,2,1,""],find_dictionaly:[4,2,1,""],font_dialog:[4,2,1,""],font_size_Change:[4,2,1,""],frame:[4,2,1,""],frame_character:[4,2,1,""],frame_image:[4,2,1,""],get_current_insert_word:[4,2,1,""],get_keywords:[4,2,1,""],initialize:[4,2,1,""],is_hiragana:[4,2,1,""],line_highlight:[4,2,1,""],message_window:[4,2,1,""],moji_count:[4,2,1,""],mouse_image_scroll:[4,2,1,""],mouse_y_scroll:[4,2,1,""],new_file:[4,2,1,""],new_open:[4,2,1,""],on_double_click:[4,2,1,""],on_double_click_yahoo:[4,2,1,""],on_name_click:[4,2,1,""],open_file:[4,2,1,""],open_file_save:[4,2,1,""],open_help:[4,2,1,""],open_url:[4,2,1,""],open_version:[4,2,1,""],overwrite_save_file:[4,2,1,""],paste:[4,2,1,""],path_read_image:[4,2,1,""],path_read_text:[4,2,1,""],print_gif:[4,2,1,""],push_keys:[4,2,1,""],pyttsx3_onend:[4,2,1,""],pyttsx3_onreadend:[4,2,1,""],pyttsx3_onword:[4,2,1,""],read_text:[4,2,1,""],redo:[4,2,1,""],remove_list:[4,2,1,""],replacement:[4,2,1,""],replacement_dialog:[4,2,1,""],replacement_dialog_on_closing:[4,2,1,""],resize_gif:[4,2,1,""],ruby:[4,2,1,""],save_charactor_file:[4,2,1,""],save_file:[4,2,1,""],search:[4,2,1,""],search_forward:[4,2,1,""],search_next:[4,2,1,""],selection:[4,2,1,""],tab:[4,2,1,""],tree_get_loop:[4,2,1,""],undo:[4,2,1,""],update_line_numbers:[4,2,1,""],yahoo:[4,2,1,""],yahoocall:[4,2,1,""],yahooresult:[4,2,1,""]},"processingmenu.ProcessingMenuClass":{find_wikipedia:[5,2,1,""],open_becoming_novelist_page:[5,2,1,""]},helpmenu:{HelpMenuClass:[0,1,1,""]},mydialog:{Mydialog:[3,1,1,""]},novel_editor:{CustomText:[4,1,1,""],LineFrame:[4,1,1,""],on_closing:[4,4,1,""]},processingmenu:{ProcessingMenuClass:[5,1,1,""]}},objnames:{"0":["py","module","Python \u30e2\u30b8\u30e5\u30fc\u30eb"],"1":["py","class","Python \u30af\u30e9\u30b9"],"2":["py","method","Python \u30e1\u30bd\u30c3\u30c9"],"3":["py","attribute","Python \u306e\u5c5e\u6027"],"4":["py","function","Python \u306e\u95a2\u6570"]},objtypes:{"0":"py:module","1":"py:class","2":"py:method","3":"py:attribute","4":"py:function"},terms:{"(%":4,"(http":10,"(true":4,"(wikipedia":10,")<":6,")\u3001":9,"--":[0,3,4,5],"-api":[5,10],".bat":8,".event":[3,4],".frame":4,".html":0,".ned":9,".py":7,".py\u30d5\u30a1\u30a4\u30eb":10,".text":4,".ttk":4,"/novel":7,"/py":8,"2exe":8,"</":6,"<<":4,"==":[8,10],"=\"":6,">>":4,">\n<":6,"?xml":6,"\"?":6,"\u3001.":9,"\u3001<":4,"\u3002(":[9,10],"\u3002)":[9,10],"\u3042\u308a":9,"\u3042\u308b":9,"\u3042\u308c":4,"\u3044\u308b":[3,4,6],"\u304a\u3053\u306a\u3046":4,"\u304a\u9858\u3044":10,"\u304b\u304b\u308b":4,"\u304b\u3068":10,"\u304b\u3089":[4,5],"\u304c\u308f\u304b\u308b":4,"\u304f\u3060":[8,10],"\u3053\u3068":[4,10],"\u3053\u308c":10,"\u3055\u3044":[8,10],"\u3059\u3079\u3066":4,"\u3059\u308b":[0,3,4,5,7,8,9],"\u3059\u308c":[4,10],"\u305d\u3053\u306b":9,"\u305d\u306e":[4,6,7],"\u305d\u308c":10,"\u305d\u308c\u305e\u308c":[4,9],"\u305f\u3044":4,"\u305f\u3081":[4,7,8],"\u305f\u3089":4,"\u3060\u3051":[4,9],"\u3064\u304e":7,"\u3064\u3051":[4,9],"\u3066\u308b":4,"\u3067\u304d":[4,9,10],"\u3067\u304d\u308b":10,"\u3067\u304d\u308c":4,"\u3067\u3059":[8,9,10],"\u3068\u304d":[3,4,9],"\u3069\u3046":4,"\u306a\u3044":[0,4,10],"\u306a\u304b":4,"\u306a\u3051\u308c":4,"\u306a\u306a\u3089":4,"\u306a\u3089":4,"\u306a\u308d":[4,5],"\u306b\u3064\u3044\u3066":1,"\u306b\u3088\u3063\u3066":9,"\u306e\u3067":[8,9,10],"\u306e\u307f":4,"\u306f\u305a":10,"\u3072\u3089":4,"\u3075\u308b":4,"\u307e\u3057":10,"\u307e\u3059":[7,8,9,10],"\u307e\u305b":10,"\u307e\u3067":[0,4],"\u3082\u306e":4,"\u3084\u3081":4,"\u3088\u3046":[0,4],"\u3089\u308c":4,"\u308c\u308b":4,"\u30a2\u30a4\u30c6\u30e0":4,"\u30a4\u30d9\u30f3\u30c8":[4,9],"\u30a4\u30e1\u30fc\u30b8":[4,9],"\u30a4\u30e1\u30fc\u30b8\u30a4\u30d9\u30f3\u30c8":4,"\u30a4\u30e1\u30fc\u30b8\u30a4\u30f3\u30b9\u30bf\u30f3\u30b9":4,"\u30a4\u30e1\u30fc\u30b8\u30ad\u30e3\u30f3\u30d0\u30b9":4,"\u30a4\u30e1\u30fc\u30b8\u30d5\u30a1\u30a4\u30eb":[0,4],"\u30a4\u30e1\u30fc\u30b8\u30d5\u30ec\u30fc\u30e0":4,"\u30a4\u30f3\u30b9\u30bf\u30f3\u30b9":[3,4,5],"\u30a4\u30f3\u30b9\u30c8\u30fc\u30eb":[7,8,10],"\u30a4\u30f3\u30d7\u30c3\u30c8\u30dc\u30c3\u30af\u30b9":3,"\u30a4\u30f3\u30d7\u30c3\u30c8\u30dc\u30c3\u30af\u30b9\u30a4\u30f3\u30b9\u30bf\u30f3\u30b9":3,"\u30a6\u30a3\u30b8\u30a7\u30c3\u30c8":4,"\u30a6\u30a4\u30f3\u30c9\u30a6":4,"\u30a6\u30a4\u30f3\u30c9\u30a6\u30a4\u30f3\u30b9\u30bf\u30f3\u30b9":0,"\u30a6\u30a7\u30b8\u30c3\u30c8":4,"\u30a8\u30f3\u30b8\u30f3":10,"\u30ab\u30a6\u30f3\u30c8":4,"\u30ab\u30d0\u30ec\u30c3\u30b8":1,"\u30ab\u30d0\u30ec\u30c3\u30b8\u30ec\u30dd\u30fc\u30c8":7,"\u30ad\u30e3\u30e9\u30af\u30bf\u30fc":[4,9],"\u30ad\u30e3\u30e9\u30af\u30bf\u30fc\u30d5\u30a1\u30a4\u30eb":[1,4],"\u30ad\u30e3\u30f3\u30d0\u30b9":4,"\u30ad\u30fc":4,"\u30ad\u30fc\u30ef\u30fc\u30c9":4,"\u30af\u30e9\u30b9":3,"\u30af\u30ea\u30c3\u30af":[3,4],"\u30b3\u30de\u30f3\u30c9":7,"\u30b3\u30fc\u30c9":[7,8,10],"\u30b5\u30a4\u30ba":4,"\u30b9\u30af\u30ed\u30fc\u30eb":4,"\u30b9\u30af\u30ed\u30fc\u30eb\u30d0\u30fc":4,"\u30bb\u30fc\u30d6\u30d5\u30a1\u30a4\u30eb":6,"\u30bd\u30d5\u30c8":9,"\u30bd\u30fc\u30b9":[0,3,4,5],"\u30bf\u30a4\u30c8\u30eb":4,"\u30bf\u30b0":4,"\u30bf\u30d6":4,"\u30bf\u30fc\u30df\u30ca\u30eb\u30a8\u30c7\u30a3\u30bf":[8,10],"\u30c0\u30a4\u30a2\u30ed\u30b0":[0,3,4],"\u30c0\u30a4\u30a2\u30ed\u30b0\u30a6\u30a4\u30f3\u30c9\u30a6\u30a4\u30f3\u30b9\u30bf\u30f3\u30b9":3,"\u30c0\u30a4\u30a2\u30ed\u30b0\u30dc\u30bf\u30f3\u30af\u30ea\u30c3\u30af":3,"\u30c0\u30d6\u30eb\u30af\u30ea\u30c3\u30af":4,"\u30c4\u30ea\u30fc\u30d3\u30e5\u30fc":4,"\u30c4\u30ea\u30fc\u30d3\u30e5\u30fc\u30a4\u30d9\u30f3\u30c8":4,"\u30c6\u30ad\u30b9\u30c8":[4,9],"\u30c6\u30ad\u30b9\u30c8\u30a4\u30d9\u30f3\u30c8":4,"\u30c6\u30ad\u30b9\u30c8\u30d5\u30a1\u30a4\u30eb":[4,9],"\u30c6\u30ad\u30b9\u30c8\u30d5\u30a9\u30eb\u30c0":9,"\u30c6\u30ad\u30b9\u30c8\u30dc\u30c3\u30af\u30b9":4,"\u30c7\u30fc\u30bf":[4,9],"\u30c8\u30c3\u30d7\u30ec\u30d9\u30eb":0,"\u30cf\u30a4\u30e9\u30a4\u30c8":4,"\u30d0\u30c3\u30c1\u30d5\u30a1\u30a4\u30eb":8,"\u30d0\u30fc\u30b8\u30e7\u30f3":[0,10],"\u30d1\u30b9":4,"\u30d1\u30e9\u30e1\u30fc\u30bf":[0,3,4,5],"\u30d5\u30a1\u30a4\u30eb":[1,4,6],"\u30d5\u30a9\u30eb\u30c0":[6,7,8,9,10],"\u30d5\u30a9\u30f3\u30c8":4,"\u30d5\u30a9\u30f3\u30c8\u30b5\u30a4\u30ba\u30c0\u30a4\u30a2\u30ed\u30b0":4,"\u30d5\u30a9\u30fc\u30ab\u30b9":4,"\u30d5\u30ec\u30fc\u30e0":4,"\u30d9\u30fc\u30b9\u30af\u30e9\u30b9":[0,3,4,5],"\u30da\u30fc\u30b8":0,"\u30dc\u30bf\u30f3":[3,4],"\u30dc\u30c3\u30af\u30b9":4,"\u30de\u30a6\u30b9\u30db\u30a4\u30fc\u30eb":4,"\u30e1\u30a4\u30f3":4,"\u30e1\u30a4\u30f3\u30bd\u30fc\u30b9\u30d5\u30a1\u30a4\u30eb":10,"\u30e1\u30a4\u30f3\u30d5\u30ec\u30fc\u30e0":4,"\u30e6\u30fc\u30b6\u30fc\u30da\u30fc\u30b8":5,"\u30e9\u30a4\u30d6\u30e9\u30ea":[4,5,10],"\u30ea\u30b5\u30a4\u30ba":4,"\u30ea\u30b9\u30c8":4,"\u30eb\u30d3":4,"\u30fb\u305d\u308c\u305e\u308c":4,"\u30fb\u30a4\u30e1\u30fc\u30b8\u30ad\u30e3\u30f3\u30d0\u30b9":4,"\u30fb\u30a4\u30e1\u30fc\u30b8\u30d5\u30a1\u30a4\u30eb":4,"\u30fb\u30a4\u30f3\u30bf\u30fc\u30cd\u30c3\u30c8\u30d6\u30e9\u30a6\u30b6":5,"\u30fb\u30a6\u30a8\u30d6\u30d6\u30e9\u30a6\u30b6":0,"\u30fb\u30ad\u30e3\u30e9\u30af\u30bf\u30fc":4,"\u30fb\u30b5\u30a4\u30ba":4,"\u30fb\u30bd\u30d5\u30c8":4,"\u30fb\u30bf\u30d6\u30ad\u30fc":4,"\u30fb\u30c4\u30ea\u30fc\u30d3\u30e5\u30fc":4,"\u30fb\u30c6\u30ad\u30b9\u30c8":4,"\u30fb\u30c6\u30ad\u30b9\u30c8\u30dc\u30c3\u30af\u30b9":4,"\u30fb\u30d0\u30fc\u30b8\u30e7\u30f3":0,"\u30fb\u30d1\u30b9":4,"\u30fb\u30d5\u30a1\u30a4\u30eb":4,"\u30fb\u30d5\u30a9\u30f3\u30c8\u30b5\u30a4\u30ba\u30c0\u30a4\u30a2\u30ed\u30b0":4,"\u30fb\u30e1\u30a4\u30f3":4,"\u30fb\u30e1\u30a4\u30f3\u30a6\u30a4\u30f3\u30c9\u30a6":4,"\u30fb\u30ea\u30b9\u30c8\u30dc\u30c3\u30af\u30b9":4,"\u30fb\u4e0e\u3048":4,"\u30fb\u4f3c\u9854\u7d75":4,"\u30fb\u4f55":4,"\u30fb\u4fdd\u5b58":4,"\u30fb\u5165\u529b\u7b49":4,"\u30fb\u5909\u6570":4,"\u30fb\u5b50\u30a2\u30a4\u30c6\u30e0":4,"\u30fb\u6587\u5b57\u6570":4,"\u30fb\u6587\u7ae0\u8aad":4,"\u30fb\u6607\u9806":4,"\u30fb\u6821\u6b63":4,"\u30fb\u691c\u7d22":4,"\u30fb\u6d88\u53bb":4,"\u30fb\u73fe\u5728":4,"\u30fb\u7f6e\u63db":4,"\u30fb\u81ea\u4f5c\u30c0\u30a4\u30a2\u30ed\u30b0":3,"\u30fb\u884c\u756a\u53f7":4,"\u30fb\u88dc\u5b8c":4,"\u30fb\u958b\u3044":4,"\u30fb\u958b\u304f":4,"\u30fbcopy":4,"\u30fbcut":4,"\u30fbpaste":4,"\u30fbpyttsx":4,"\u30fbredo":4,"\u30fbtext":4,"\u30fbuedo":4,"\u30fbwikipedia":5,"\u30fbyahoo":4,"\u4e0a\u3052":4,"\u4e0a\u66f8":4,"\u4e0a\u66f8\u304d":4,"\u4e0a\u8a18":10,"\u4e0b\u8a18":[8,10],"\u4e0d\u660e":4,"\u4e21\u65b9":4,"\u4e2d\u65ad":4,"\u4e2d\u6b62":4,"\u4ed8\u3051":4,"\u4ed8\u3051\u308b":4,"\u4ee5\u5916":10,"\u4f3c\u9854":[4,9],"\u4f3c\u9854\u7d75":4,"\u4f46\u3057":[9,10],"\u4f4d\u7f6e":4,"\u4f5c\u6210":[1,3,4],"\u4f7f\u3044":8,"\u4f7f\u3046":9,"\u4f7f\u3063":[0,4,5,10],"\u4f7f\u7528":[1,7,8],"\u4fdd\u5b58":[1,4,6],"\u5019\u88dc":4,"\u5165\u3063":[6,10],"\u5165\u529b":[3,4,7],"\u51e6\u7406":[3,4,10],"\u51fa\u3057":4,"\u51fa\u3059":4,"\u51fa\u529b":7,"\u521d\u3081":4,"\u521d\u671f":4,"\u521d\u671f\u5316":4,"\u5224\u65ad":4,"\u5236\u4f5c":4,"\u524a\u9664":4,"\u5316\u3051":9,"\u5316\u53ca":4,"\u5358\u8a9e":4,"\u53d6\u308c":10,"\u53d6\u5f97":[3,4],"\u53f3\u5074":4,"\u5408\u3048":10,"\u5408\u6210":10,"\u540c\u3058":[4,9],"\u540d\u524d":[4,9],"\u547c\u3073\u51fa\u3057":[3,4],"\u547c\u3073\u51fa\u3059":4,"\u547c\u3073\u540d":[4,6],"\u56de\u3057":4,"\u5834\u6240":[4,9],"\u5909\u3048\u308b":4,"\u5909\u63db":[8,10],"\u5909\u66f4":[4,9],"\u5927\u9805\u76ee":9,"\u59cb\u3081\u308b":4,"\u5b58\u5728":4,"\u5b9f\u4f53":9,"\u5b9f\u884c":[1,7,10],"\u5c0b\u306d":4,"\u5c0f\u8aac":9,"\u5c0f\u8aac\u5bb6":[4,5],"\u5c0f\u9805\u76ee":9,"\u5f35\u308a":4,"\u5f62\u5f0f":[4,6,9],"\u5f62\u614b":[4,10],"\u5f8c\u4ee5":7,"\u5f97\u308b":4,"\u601d\u3044":10,"\u6027\u5225":6,"\u6050\u308c":9,"\u60c5\u5831":0,"\u610f\u5473":5,"\u623b\u3059":[4,9],"\u623b\u308a\u5024":[3,4],"\u62bc\u3055":4,"\u62bc\u3057":4,"\u62bc\u3059":0,"\u62bc\u4e0b":4,"\u62e1\u5927":[4,9],"\u62e1\u5927\u7387":4,"\u62e1\u5f35":4,"\u62e1\u5f35\u5b50":9,"\u632f\u308b":4,"\u633f\u5165":4,"\u63cf\u753b":4,"\u652f\u63f4":4,"\u6570\u5b57":9,"\u6587\u5b57":[4,5,9,10],"\u6587\u5b57\u5217":4,"\u6587\u5b57\u6570":4,"\u6587\u7ae0":[4,10],"\u6587\u7ae0\u8aad":4,"\u65b0\u898f":4,"\u65b9\u6cd5":[1,7],"\u6607\u9806":4,"\u66f8\u304b":4,"\u6700\u4e2d":4,"\u6700\u521d":4,"\u672c\u540d":6,"\u6821\u6b63":4,"\u691c\u7d22":[4,5,10],"\u6a19\u6e96":10,"\u6b62\u3081\u308b":4,"\u6d88\u3048":0,"\u6d88\u53bb":4,"\u6e96\u5099":4,"\u72b6\u614b":4,"\u73fe\u5728":4,"\u73fe\u5728\u884c":4,"\u74b0\u5883":10,"\u753b\u50cf":[4,9,10],"\u753b\u9762":4,"\u7565\u6b74":[4,6],"\u76f8\u5bfe":4,"\u78ba\u8a8d":[4,10],"\u79fb\u52d5":4,"\u7a7a\u6b04":4,"\u7b87\u6240":4,"\u7d20\u89e3":[4,10],"\u7d22\u5f15":1,"\u7d42\u3048":4,"\u7d42\u308f\u3063":4,"\u7d42\u4e86":4,"\u7d50\u679c":4,"\u7d9a\u3051\u308b":4,"\u7e2e\u5c0f":4,"\u7e70\u308a\u8fd4\u3057":4,"\u7f6e\u304f":9,"\u7f6e\u63db":4,"\u8077\u7a2e":9,"\u81ea\u5df1":10,"\u8272\u4ed8\u304d":4,"\u884c\u3046":4,"\u884c\u6570":4,"\u884c\u756a":4,"\u884c\u756a\u53f7":4,"\u8868\u793a":[0,3,4,9],"\u88dc\u5b8c":4,"\u89e3\u51cd":9,"\u8a08\u6e2c":1,"\u8a18\u5165":[8,10],"\u8a2d\u5b9a":4,"\u8a72\u5f53":4,"\u8a73\u7d30":7,"\u8a95\u751f":[4,6],"\u8aad\u307f":4,"\u8aad\u307f\u4e0a\u3052":[4,10],"\u8aad\u307f\u4e0a\u3052\u308b":4,"\u8aad\u307f\u8fbc\u3093":4,"\u8cac\u4efb":10,"\u8cbc\u308a\u4ed8\u3051\u308b":4,"\u8d77\u52d5":[4,8,10],"\u8ffd\u52a0":4,"\u9055\u3046":4,"\u9078\u629e":[4,5],"\u914d\u7f6e":4,"\u9577\u3044":4,"\u9589\u3058":4,"\u9589\u3058\u308b":4,"\u958b\u3044":4,"\u958b\u304f":[0,4,5],"\u958b\u767a":10,"\u95a2\u4fc2":4,"\u95a2\u9023":4,"\u97f3\u58f0":10,"\u9805\u76ee":4,"case":4,"char":4,"class":[0,3,4,5],"false":4,"int":4,"new":4,"true":4,_becoming:5,_btn:4,_change:4,_character:4,_charactor:4,_click:4,_closing:4,_complete:4,_count:4,_current:4,_data:0,_dialog:4,_dictionaly:4,_double:4,_editor:[1,2,7],_event:4,_file:4,_forward:4,_get:4,_gif:4,_help:4,_highlight:4,_hiragana:4,_image:4,_insert:4,_keys:4,_keywords:4,_line:4,_list:4,_loop:4,_name:[3,4],_next:4,_novelist:5,_numbers:4,_ok:3,_on:4,_open:4,_page:5,_path:4,_read:4,_save:4,_scroll:4,_setting:4,_size:4,_tags:4,_text:4,_url:4,_version:4,_widgets:4,_wiki:5,_wikipedia:5,_win:3,_window:4,_word:4,_y:4,_yahoo:4,all:4,app:5,appid:4,auto:4,birthday:6,bit:10,body:6,bool:4,btn:4,button:3,call:6,change:4,character:[6,9],clear:4,client:4,completed:4,contents:1,copy:4,coverage:7,coveragepy:7,create:4,ctrl:4,customtext:4,cut:4,data:[6,9],dist:8,editor:10,event:[3,4,9],ex:4,exe:1,externalloop:4,find:[4,5],font:4,frame:4,get:4,gif:[4,9],help:0,helpmenu:[1,2],helpmenuclass:0,html:[4,7],htmlcov:7,id:4,im:4,image:[0,4,9],index:4,initialize:4,install:[7,8,10],instance:[0,3,4,5],is:4,jaconv:10,janome:[4,10],kwargs:4,length:4,line:4,lineframe:4,location:4,master:4,message:[3,4],module:[1,2],moji:4,mouse:4,mydialog:[1,2],name:[4,6],nobel:9,none:[3,4],novel:[2,10],object:[0,3,5],occupation:9,on:4,onend:4,onreadend:4,onword:4,open:[4,5],overwrite:4,paste:4,path:4,pillow:10,pip:[7,8,10],print:4,processingmenu:[1,2],processingmenuclass:5,pt:4,push:4,px:4,py:8,pyinstaller:8,python:[7,10],pyttsx:[4,10],read:4,readme:0,redo:4,remove:4,replacement:4,requests:10,resize:4,root:0,ruby:4,run:7,save:4,scale:4,scroll:4,search:4,selection:4,self:3,sentence:4,sex:6,source:[1,7,10],space:9,str:[0,3,4],sub:3,tab:4,text:[3,4],title:[3,4],tkinter:[3,4],tree:4,txt:3,type:3,ubuntu:10,undo:4,update:4,version:[0,6,10],wiki:5,wikipedia:[5,10],windows:[8,10],xml:4,yahoo:4,yahoocall:4,yahooresult:4,zip:9},titles:["helpmenu module","Welcome to Novel Editor's documentation!","source","mydialog module","novel_editor module","processingmenu module","\u30ad\u30e3\u30e9\u30af\u30bf\u30fc\u30d5\u30a1\u30a4\u30eb\u306b\u3064\u3044\u3066","\u30ab\u30d0\u30ec\u30c3\u30b8\u306e\u8a08\u6e2c\u306b\u3064\u3044\u3066","exe\u5b9f\u884c\u30d5\u30a1\u30a4\u30eb\u3092\u4f5c\u6210","\u4fdd\u5b58\u30d5\u30a1\u30a4\u30eb\u306b\u3064\u3044\u3066","\u4f7f\u7528\u65b9\u6cd5"],titleterms:{"'s":1,"\u306b\u3064\u3044\u3066":[6,7,9],"\u30ab\u30d0\u30ec\u30c3\u30b8":7,"\u30ad\u30e3\u30e9\u30af\u30bf\u30fc\u30d5\u30a1\u30a4\u30eb":6,"\u30d5\u30a1\u30a4\u30eb":[8,9],"\u4f5c\u6210":8,"\u4f7f\u7528":10,"\u4fdd\u5b58":9,"\u5b9f\u884c":8,"\u65b9\u6cd5":10,"\u8a08\u6e2c":7,_editor:4,and:1,documentation:1,editor:1,exe:8,helpmenu:0,indices:1,module:[0,3,4,5],mydialog:3,novel:[1,4],processingmenu:5,source:2,tables:1,to:1,welcome:1}})
