+++ docs/html/sub/coverage.html
@@ -37,7 +37,7 @@
     
     <link rel="index" title="索引" href="../genindex.html" />
     <link rel="search" title="検索" href="../search.html" />
-    <link rel="next" title="novel_editor モジュール" href="../novel_editor.html" />
+    <link rel="next" title="source" href="../modules.html" />
     <link rel="prev" title="exe実行ファイルを作成" href="exe.html" /> 
 </head>
 
@@ -90,7 +90,7 @@
 <li class="toctree-l1"><a class="reference internal" href="save.html">保存ファイルについて</a></li>
 <li class="toctree-l1"><a class="reference internal" href="exe.html">exe実行ファイルを作成</a></li>
 <li class="toctree-l1 current"><a class="current reference internal" href="#">カバレッジの計測について</a></li>
-<li class="toctree-l1"><a class="reference internal" href="../novel_editor.html">novel_editor モジュール</a></li>
+<li class="toctree-l1"><a class="reference internal" href="../modules.html">source</a></li>
 </ul>
 
             
@@ -186,7 +186,7 @@
   
     <div class="rst-footer-buttons" role="navigation" aria-label="footer navigation">
       
-        <a href="../novel_editor.html" class="btn btn-neutral float-right" title="novel_editor モジュール" accesskey="n" rel="next">Next <span class="fa fa-arrow-circle-right"></span></a>
+        <a href="../modules.html" class="btn btn-neutral float-right" title="source" accesskey="n" rel="next">Next <span class="fa fa-arrow-circle-right"></span></a>
       
       
         <a href="exe.html" class="btn btn-neutral float-left" title="exe実行ファイルを作成" accesskey="p" rel="prev"><span class="fa fa-arrow-circle-left"></span> Previous</a>
