+++ sphinx/helpmenu.rst
@@ -1,0 +1,7 @@
+helpmenu module
+===============
+
+.. automodule:: helpmenu
+   :members:
+   :undoc-members:
+   :show-inheritance:
