+++ sphinx/modules.rst
@@ -4,4 +4,7 @@ source
 .. toctree::
    :maxdepth: 4
 
+   helpmenu
+   mydialog
    novel_editor
+   processingmenu
