+++ sphinx/mydialog.rst
@@ -1,0 +1,7 @@
+mydialog module
+===============
+
+.. automodule:: mydialog
+   :members:
+   :undoc-members:
+   :show-inheritance:
