+++ sphinx/processingmenu.rst
@@ -1,0 +1,7 @@
+processingmenu module
+=====================
+
+.. automodule:: processingmenu
+   :members:
+   :undoc-members:
+   :show-inheritance:
